pyproject.toml
[project]
name = "kernel-debug"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.16.0",
    "pymupdf>=1.26.5",
    "streamlit>=1.50.0",
]
